Move files related to MediaService to service package
